Refactor 670.
